backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
faiss-cpu==1.7.4
gunicorn==21.2.0